+++ modules/project/import-core/src/sap/commerce/toolset/project/configurator/ApplicationSettingsConfigurator.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -33,6 +33,7 @@ class ApplicationSettingsConfigurator : ProjectImportConfigurator {
             this.externalDbDriversDirectory = context.externalDbDriversDirectory?.toSystemIndependentName
 
             this.importOOTBModulesInReadOnlyMode = importSettings.importOOTBModulesInReadOnlyMode
+            this.useRealPathForModuleRoot = importSettings.useRealPathForModuleRoot
             this.importCustomAntBuildFiles = importSettings.importCustomAntBuildFiles
             this.ignoreNonExistingSourceDirectories = importSettings.ignoreNonExistingSourceDirectories
             this.hideEmptyMiddleFolders = importSettings.hideEmptyMiddleFolders