+++ modules/project/import-core/src/sap/commerce/toolset/project/configurator/ProjectSettingsConfigurator.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -53,6 +53,7 @@ class ProjectSettingsConfigurator : ProjectImportConfigurator {
         val importSettings = context.settings
 
         projectSettings.importOotbModulesInReadOnlyMode = importSettings.importOOTBModulesInReadOnlyMode
+        projectSettings.useRealPathForModuleRoot = importSettings.useRealPathForModuleRoot
         projectSettings.importCustomAntBuildFiles = importSettings.importCustomAntBuildFiles
         projectSettings.useFakeOutputPathForCustomExtensions = importSettings.useFakeOutputPathForCustomExtensions
         projectSettings.withDecompiledOotbSources = importSettings.withDecompiledOotbSources