+++ modules/shared/core/src/sap/commerce/toolset/settings/ApplicationSettings.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -53,6 +53,11 @@ class ApplicationSettings : SerializablePersistentStateComponent<ApplicationSett
         get() = state.defaultPlatformInReadOnly
         set(value) {
             updateState { it.copy(defaultPlatformInReadOnly = value) }
+        }
+    var useRealPathForModuleRoot: Boolean
+        get() = state.useRealPathForModuleRoot
+        set(value) {
+            updateState { it.copy(useRealPathForModuleRoot = value) }
         }
     var ignoreNonExistingSourceDirectories: Boolean
         get() = state.ignoreNonExistingSourceDirectories