+++ modules/shared/core/src/sap/commerce/toolset/settings/state/ApplicationSettingsState.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -30,6 +30,7 @@ data class ApplicationSettingsState(
     @JvmField @OptionTag val groupExternalModules: Boolean = false,
     @JvmField @OptionTag val hideEmptyMiddleFolders: Boolean = true,
     @JvmField @OptionTag val defaultPlatformInReadOnly: Boolean = true,
+    @JvmField @OptionTag val useRealPathForModuleRoot: Boolean = false,
     @JvmField @OptionTag val ignoreNonExistingSourceDirectories: Boolean = false,
     @JvmField @OptionTag val librarySourcesFetchMode: LibrarySourcesFetchMode = LibrarySourcesFetchMode.REMOTE,
     @JvmField @OptionTag val withExternalLibrarySources: Boolean = true,